+++ sdk/documentintelligence/azure-ai-documentintelligence/src/test/java/com/azure/ai/documentintelligence/TestUtils.java
@@ -50,6 +50,7 @@ public final class TestUtils {
     static final String W2_JPG = "w2-single.png";
     static final String IRS_1040 = "IRS-1040_3.pdf";
     static final String LAYOUT_SAMPLE = "layout-pageobject.pdf";
+    static final String BATCH_SAMPLE_PDF = "Acord_27.pdf";
 
     static final String EXPECTED_MERCHANT_NAME = "Contoso";
     static final String URL_TEST_FILE_FORMAT = "https://raw.githubusercontent.com/Azure/azure-sdk-for-java/"
@@ -72,6 +73,8 @@ public final class TestUtils {
         GLOBAL_CONFIGURATION.get("DOCUMENTINTELLIGENCE_SELECTION_MARK_DATA_CONTAINER_SAS_URL");
     public static final String DOCUMENTINTELLIGENCE_CLASSIFIER_TRAINING_DATA_CONTAINER_SAS_URL_CONFIGURATION =
         GLOBAL_CONFIGURATION.get("DOCUMENTINTELLIGENCE_CLASSIFIER_TRAINING_DATA_CONTAINER_SAS_URL");
+    public static final String DOCUMENTINTELLIGENCE_BATCH_TRAINING_DATA_CONTAINER_SAS_URL_CONFIGURATION =
+        GLOBAL_CONFIGURATION.get("DOCUMENTINTELLIGENCE_BATCH_TRAINING_DATA_CONTAINER_SAS_URL");
     public static final Duration DEFAULT_POLL_INTERVAL = Duration.ofSeconds(5);
     public static final Duration DEFAULT_TIMEOUT = Duration.ofSeconds(30);
     private TestUtils() {
@@ -104,7 +107,9 @@ public final class TestUtils {
     public static void getClassifierTrainingDataContainerHelper(Consumer<String> testRunner, boolean isPlaybackMode) {
         testRunner.accept(getClassifierTrainingFilesContainerUrl(isPlaybackMode));
     }
-
+    public static void getBatchTrainingDataContainerHelper(Consumer<String> testRunner, boolean isPlaybackMode) {
+        testRunner.accept(getBatchTrainingFilesContainerUrl(isPlaybackMode));
+    }
     /**
      * Get the testing data set SAS Url value based on the test running mode.
      *
@@ -151,6 +156,15 @@ public final class TestUtils {
      */
     private static String getClassifierTrainingFilesContainerUrl(boolean isPlaybackMode) {
         return isPlaybackMode ? "https://isPlaybackmode" : DOCUMENTINTELLIGENCE_CLASSIFIER_TRAINING_DATA_CONTAINER_SAS_URL_CONFIGURATION;
+    }
+
+    /**
+     * Get the training data set SAS Url value based on the test running mode.
+     *
+     * @return the training data set Url
+     */
+    private static String getBatchTrainingFilesContainerUrl(boolean isPlaybackMode) {
+        return isPlaybackMode ? "https://isPlaybackmode" : DOCUMENTINTELLIGENCE_BATCH_TRAINING_DATA_CONTAINER_SAS_URL_CONFIGURATION;
     }
 
     /**