+++ sdk/resourcemanager/azure-resourcemanager-compute/src/test/java/com/azure/resourcemanager/compute/VirtualMachineExtensionOperationsTests.java
@@ -13,12 +13,19 @@ import com.azure.resourcemanager.compute.models.KnownLinuxVirtualMachineImage;
 import com.azure.resourcemanager.compute.models.VirtualMachine;
 import com.azure.resourcemanager.compute.models.VirtualMachineExtension;
 import com.azure.resourcemanager.compute.models.VirtualMachineSizeTypes;
+import com.azure.resourcemanager.keyvault.models.Secret;
+import com.azure.resourcemanager.keyvault.models.Vault;
 import com.azure.resourcemanager.storage.models.StorageAccount;
+import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.io.IOException;
 import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.Base64;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 
 public class VirtualMachineExtensionOperationsTests extends ComputeManagementTest {
@@ -321,5 +328,59 @@ public class VirtualMachineExtensionOperationsTests extends ComputeManagementTes
         // In deallocated state, we can get VM's extensions but not their instance views.
         Assertions.assertTrue(vm.listExtensions().size() > 0);
         Assertions.assertTrue(vm.listExtensions().values().stream().allMatch(extension -> extension.getInstanceView() == null));
+    }
+
+    @Test
+    public void canIgnoreInvalidJson() throws IOException {
+        String vmName = generateRandomResourceName("javavm", 15);
+        String vaultName = generateRandomResourceName("javavt", 15);
+        final String secretName = generateRandomResourceName("srt", 10);
+
+        Vault vault =
+            this
+                .keyVaultManager
+                .vaults()
+                .define(vaultName)
+                .withRegion(region)
+                .withNewResourceGroup(rgName)
+                .defineAccessPolicy()
+                .forUser(azureCliSignedInUser().userPrincipalName())
+                .allowSecretAllPermissions()
+                .attach()
+                .withDeploymentEnabled()
+                .create();
+        final InputStream embeddedJsonConfig =
+            this.getClass().getResourceAsStream("/myTest.txt");
+        String secretValue = IOUtils.toString(embeddedJsonConfig, StandardCharsets.UTF_8);
+        Secret secret = vault.secrets().define(secretName).withValue(secretValue).create();
+
+        Map<String, Object> extensionSecretSettings = new HashMap<>();
+        extensionSecretSettings.put("pollingIntervalInS", "3600");
+        extensionSecretSettings.put("observedCertificates", Collections.singletonList(secret.id()));
+
+        VirtualMachine vm =
+            computeManager
+                .virtualMachines()
+                .define(vmName)
+                .withRegion(region)
+                .withExistingResourceGroup(rgName)
+                .withNewPrimaryNetwork("10.0.0.0/28")
+                .withPrimaryPrivateIPAddressDynamic()
+                .withoutPrimaryPublicIPAddress()
+                .withPopularLinuxImage(KnownLinuxVirtualMachineImage.UBUNTU_SERVER_18_04_LTS)
+                .withRootUsername("Foo12")
+                .withSsh(sshPublicKey())
+                .withSize(VirtualMachineSizeTypes.fromString("Standard_D2a_v4"))
+                .defineNewExtension("KeyVaultForLinux")
+                .withPublisher("Microsoft.Azure.KeyVault")
+                .withType("KeyVaultForLinux")
+                .withVersion("1.0")
+                .withPublicSetting("secretsManagementSettings", extensionSecretSettings)
+                .attach()
+                .create();
+        VirtualMachineExtension extension = vm.listExtensions().get("KeyVaultForLinux");
+        Assertions.assertNotNull(extension);
+        Assertions.assertNotNull(extension.publicSettingsAsJsonString());
+        Assertions.assertFalse(extension.publicSettings().isEmpty());
     }
 }