+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Features Added
 
-### Breaking Changes
+- Supported enabling ultra SSD in `VirtualMachine` class.
 
 ### Bugs Fixed
+
+- Fixed a bug that `ClassCastException` occurs for some Virtual Machines with extensions.
 
 ### Other Changes